feat: implement AssistantComposer component with @-mention context binding and model selection
